+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ### Feature
 - Add prop `disabled` to BugDate
 - Remove prop `name` from BugRadioGroup
+- Add prop `size` to BugLoading
+- Add event `change` to BugRadioGroup and BugCheckboxGroup
 
 ## v1.1.11
 ### Feature